// model/featuresPushOptionsFcm.ts
/**
 * Whispir Platform API
 * Whispir Platform API for cross channel and multi channel communications. Documentation on each endpoint is available at https://developers.whispir.com.
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: support@whispir.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

/**
* The fcm object holds attributes specific to the contacts\' Firebase-registered Android devices
*/
export class FeaturesPushOptionsFcm {
    /**
    * Translates to the <span style=\"color:red\">*android_channel_id*</span> attribute in FCM. Note: use this attribute for devices running Android versions 8 or later.
    */
    'androidChannelId'?: string;
    /**
    * The sound to play when the device receives the notification. The value *default* will play the system sound, otherwise this attribute must point to a sound resource bundled in the app, e.g. *_/res/raw/notification_sound.wav*  Note: This attribute can be used for devices running Android versions older than Android 8. For versions 8+ please use the *androidChannelId*
    */
    'sound'?: string;
    /**
    * The action linked with a user click on the notification. Translates to the <span style=\"color:red\">*click_action*</span> attribute in FCM.
    */
    'action'?: string;
    /**
    * The URL of an image to be downloaded and displayed in the notification.  Details of the Android support for images in push notifications can be found [here](https://firebase.google.com/docs/cloud-messaging/android/send-image)
    */
    'image'?: string;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "androidChannelId",
            "baseName": "androidChannelId",
            "type": "string"
        },
        {
            "name": "sound",
            "baseName": "sound",
            "type": "string"
        },
        {
            "name": "action",
            "baseName": "action",
            "type": "string"
        },
        {
            "name": "image",
            "baseName": "image",
            "type": "string"
        }    ];

    static getAttributeTypeMap() {
        return FeaturesPushOptionsFcm.attributeTypeMap;
    }
}